move bootstrap.js to components/urifixup.js
Preparing for a non-bootstrappable add-on, because the addon's
functionality cannot be bootstrapped.
nsDocShell has a global reference to a nsIURIFixup instance, which is
initialized in the constructor. This means that replacing the
implementation of the '@mozilla.org/docshell/urifixup;1' contract does
not give the desired result (i.e. the new implementation is not used).

urifixup.js will be rewritten in later commits.