Override openLinkIn instead of nsIURIFixup
Overriding nsIURIFixup appears to be a dead end because of
https://github.com/Rob--W/https-by-default/issues/1 and
https://github.com/Rob--W/https-by-default/issues/2.

So, I no longer try to override the default nsIURIFixup implementation.
Instead, I monkey-patch the openLinkIn method in every browser window.
This method also allows me to declare the addon as restartless.

The new method is designed to be very error-tolerant, because it relies
on internal Firefox APIs, which may change at any point in the future.
The following scenarios are explicitly accounted for:
- Addon install/enable/disable/uninstall, multiple times.
- Something else also monkey-patches openLinkIn, and this addon is
  enabled/disabled before or after this external action.
- Firefox updates and removes the openLinkIn method.
- Firefox updates and changes the openLinkIn function signature.