Bump Firefox addon version to make AMO happy
While testing, I had temporarily published 0.1 of the addon. Because of
this, the last submission (the previous commit) was rejected for the
following reason:

"1) New uploads must have a higher version number than the previously
uploaded versions otherwise those files do not propagate across our
mirror network correctly, and many users will receive the old version,
which they will be unable to install due to checksum mismatches."

So the version has been bumped.

In addition to that, I received the following comment:

"Please note the following for the next update:
1) Your add-on prints debugging information to the Console, which is
generally not allowed in production add-ons."

There are only five uses of console.warn, all of them are only activated
in an extraordinary situation that most likely requires an update to the
addon (e.g. because Firefox internals change).
Besides that, the warnings are emitted only once per window, since the
functions that call console.warn are invoked once per browser window.
Therefore keeping the console.warn calls is justified.