Convert Firefox bootstrapped add-on to WebExtensions
Firefox 57 is near, and the only supported add-on type are
WebExtensions. This commit rewrites the add-on and fixes #12.

It is not possible to intercept the typed URL.
However it is possible to see whether a navigation is web-triggered or
not. So as a compromis, redirect most top-level, non-web-triggered
http:-navigations to https.

Domains without dots, IP addresses and reserved TLDs will no longer
be redirected to https. This fixes #10.