Update Chromium patch (68.0.3440.75)
Patch had to be changed because of:
https://chromium.googlesource.com/chromium/src/+/54d6ad62258f03e21bf918d647686bd7b8aa4d15

Note: This only prepends "http" to the omnibox in the default config.
If the following flag is on, then "http" is omitted (among other parts):
chrome://flags/#omnibox-ui-hide-steady-state-url-scheme-and-subdomains
To change that, the format_types in ToolbarModelImpl::GetURLForDisplay
has to be changed. See also: https://crbug.com/797354