Update Chromium patch (show http scheme)
In Chromium 68.0.3440.75, the code was refactored to hide the https
scheme by default:
https://chromium.googlesource.com/chromium/src/+/c71963abef09ea6dd2c3b53996f9580643665e5c

To avoid hiding https by default, the following flag has to be toggled:
chrome://flags/#omnibox-ui-hide-steady-state-url-scheme
This does however not work for http, which is hidden by default.
This updated patch corrects that, by showing http in the omnibox by default.

(This should have been in the patch for 68.0.3440.75,
 but now it has been tested with 73.0.3683.86)